test(header): migrate condense
